feat: :construction: Se corrigen algunos errores de sintaxis en las migraciones

+++ database/migrations/2025_06_14_030332_add_column_to_users_table.php
@@ -12,23 +12,25 @@ return new class extends Migration
     // Añadir más columnas a la tabla de usuarios
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            //Añadir columnas despues de la columna 'id()'
-            $table->foreignId('staff_id')
-                ->nullable()
-                ->constrained('staff')
-                ->onDelete('set null'); // Referencia al p ersonal asociado al usuario
-
-        });
+        // Verifica si la columna ya existe antes de agregarla
+        if (!Schema::hasColumn('users', 'staff_id')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->foreignId('staff_id')
+                    ->nullable()
+                    ->constrained('staff')
+                    ->onDelete('set null');
+            });
+        }
     }
 
     /**
      * Reverse the migrations.
      */
-    public function down(): void
+    public function down()
     {
         Schema::table('users', function (Blueprint $table) {
-            //
+            $table->dropForeign(['staff_id']);
+            $table->dropColumn('staff_id');
         });
     }
 };
